requirements.txt
# Core RAG and LLM dependencies
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.7
chromadb==0.4.22
sentence-transformers==2.2.2

# Document processing
pypdf2==3.0.1
pymupdf==1.23.21
pdf2image==1.17.0
pillow==10.2.0
pytesseract==0.3.10

# Image processing for schematics
opencv-python==4.9.0.80
numpy==1.24.3
matplotlib==3.7.4

# Web scraping and API
requests==2.31.0
beautifulsoup4==4.12.2
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.2

# CLI framework
click==8.1.7
rich==13.7.0
typer==0.9.0

# Utilities
python-dotenv==1.0.0
pathlib==1.0.1
tqdm==4.66.1